Refactor without using GenServer
Rewrote most of the game. Changing the architecture from a GenServer
to a single LiveView that shares messages through Endpoint.

As suggested [here](https://elixirforum.com/t/should-we-use-presence-to-store-player-position/67304/2?u=danielrachi).

Won't use this model because some lag is introduced and there are some
weird behaviours related to presence and the ball speed.

Resolves: na
See also: #36